+++ notes/CDEC-151/Chain/Update.hs
@@ -1,0 +1,7 @@
+module Chain.Update where
+
+import Block (Block, Point)
+
+data ChainUpdate = AddBlock Block
+                 | RollBack Point
+  deriving Show
